Move file_util module into compat
