+++ lib/fluent/plugin/file_util.rb
@@ -14,44 +14,9 @@
 #    limitations under the License.
 #
 
+require 'fluent/compat/file_util'
+
 module Fluent
-  module Plugin
-    module FileUtil
-      # Check file is writable if file exists
-      # Check directory is writable if file does not exist
-      #
-      # @param [String] path File path
-      # @return [Boolean] file is writable or not
-      def writable?(path)
-        return false if File.directory?(path)
-        return File.writable?(path) if File.exist?(path)
-
-        dirname = File.dirname(path)
-        return false if !File.directory?(dirname)
-        File.writable?(dirname)
-      end
-      module_function :writable?
-
-      # Check file is writable in conjunction wtih mkdir_p(dirname(path))
-      #
-      # @param [String] path File path
-      # @return [Boolean] file writable or not
-      def writable_p?(path)
-        return false if File.directory?(path)
-        return File.writable?(path) if File.exist?(path)
-
-        dirname = File.dirname(path)
-        until File.exist?(dirname)
-          dirname = File.dirname(dirname)
-        end
-
-        return false if !File.directory?(dirname)
-        File.writable?(dirname)
-      end
-      module_function :writable_p?
-    end
-  end
-
   # obsolete
-  FileUtil = Fluent::Plugin::FileUtil
+  FileUtil = Fluent::Compat::FileUtil
 end
